Split mask definition code from AbstractValueMapRule into another abc.
